Support Windows Containers (#870)
