+++ core/src/test/java/org/testcontainers/DockerClientFactoryTest.java
@@ -3,7 +3,6 @@ package org.testcontainers;
 import com.github.dockerjava.api.exception.NotFoundException;
 import org.junit.Test;
 import org.rnorth.visibleassertions.VisibleAssertions;
-import org.testcontainers.DockerClientFactory.DiskSpaceUsage;
 import org.testcontainers.dockerclient.LogToStringContainerCallback;
 import org.testcontainers.utility.TestcontainersConfiguration;
 
@@ -38,7 +37,7 @@ public class DockerClientFactoryTest {
     @Test
     public void shouldHandleBigDiskSize() throws Exception {
         String dfOutput = "/dev/disk1     2982480572 1491240286 2982480572    31%    /";
-        DiskSpaceUsage usage = DockerClientFactory.instance().parseAvailableDiskSpace(dfOutput);
+        DiskSpaceUsage usage = DiskSpaceUsage.parseAvailableDiskSpace(dfOutput);
 
         VisibleAssertions.assertEquals("Available MB is correct", 2982480572L / 1024L, usage.availableMB.orElse(0L));
         VisibleAssertions.assertEquals("Available percentage is correct", 31, usage.usedPercent.orElse(0));
