+++ docs/compatibility.md
@@ -9,4 +9,4 @@
 | Mac OS X - Docker Toolbox           | Docker Machine v0.8.0  |                                                                                                                                                                                                                                                                                                                                                                    |
 | Mac OS X - Docker for Mac      | 1.12.0          | *Support is best-efforts at present*. `getTestHostIpAddress()` is [not currently supported](https://github.com/testcontainers/testcontainers-java/issues/166) due to limitations in Docker for Mac. |
 | Windows - Docker Toolbox            |                             | *Support is limited at present and this is not currently tested on a regular basis*.                                                                                                                                                                                                                                                                               |
-| Windows - Docker for Windows   |                             | *Support is best-efforts at present.* Only Linux Containers (LCOW) are supported at the moment. See [Windows Support](usage/windows_support.md)                                                                                                                                                                                                                                                                                                                                        |
+| Windows - Docker for Windows   |                             | *Support is best-efforts at present.* Supported Linux containers. Windows and LCOW support is experimental. See [Windows Support](usage/windows_support.md)                                                                                                                                                                                                                                                                                                                                        |
