+++ docs/usage/windows_support.md
@@ -19,9 +19,9 @@ effort.
 ### MySQL containers
 * MySQL server prevents custom configuration file (ini-script) from being loaded due to security measures ([link to feature description](database_containers.md#using-an-init-script))
 
-### Windows Container on Windows (WCOW)
+### Windows Container
 
-* WCOW is currently not supported, since Testcontainers uses auxiliary Linux containers for certain tasks and Docker for Windows does not support hybrid engine mode at the time of writing.
+* Windows and LCOW containers support is experimental, please [report](https://github.com/testcontainers/testcontainers-java/issues) any found problems.
 
 ### Docker credential helpers
